+++ frontend/src/component/modal/contents/FriendWaitList.js
@@ -6,10 +6,9 @@ const FriendWaitList = () => {
   return `
     <div class="container p-3">
       <div class="inner-container vh-50 px-5 py-3 border border-5 border-light rounded">
-        ${waitList.map((friendNick) => FriendWaitListItem ({ nick: friendNick })).join('')}
+        ${waitList.map((friendNick) => FriendWaitListItem({ nick: friendNick })).join('')}
       </div>
     </div>
-
   `;
 };
