+++ frontend/src/component/modal/contents/RegisterFormItem.js
